requirements.txt
Flask==2.3.2               
tensorflow==2.13.0         
numpy==1.24.2              
Pillow==9.5.0              
Werkzeug==2.3.7            
Jinja2==3.1.2              
Click==8.1.7               
itsdangerous==2.1.2        
MarkupSafe==2.1.3          

